[OK] Added permission verification for user update and deletion

+++ src/modules/Users/useCases/deleteUser/DeleteUserUseCase.ts
@@ -12,11 +12,15 @@ class DeleteUserUseCase {
         private userRepository: IUserRepository
     ) { }
 
-    async execute(userId: number): Promise<IUserDTO> {
+    async execute(userId: number, permissions: string): Promise<IUserDTO> {
         const user = await this.userRepository.getById(userId);
 
         if (!user) {
             throw new AppError("Este usuário não existe", 422);
+        }
+
+        if (user.cpf !== permissions) {
+            throw new AppError("Usuário sem permissões", 403);
         }
 
         const deletedUser = await this.userRepository.deleteById(userId);
